+++ packages/block-directory/src/components/downloadable-blocks-list/index.js
@@ -2,7 +2,7 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { privateApis as componentsPrivateApis } from '@wordpress/components';
+import { Composite } from '@wordpress/components';
 import { getBlockType } from '@wordpress/blocks';
 import { useDispatch } from '@wordpress/data';
 
@@ -11,11 +11,10 @@ import { useDispatch } from '@wordpress/data';
  */
 import DownloadableBlockListItem from '../downloadable-block-list-item';
 import { store as blockDirectoryStore } from '../../store';
-import { unlock } from '../../lock-unlock';
 
-const { CompositeV2: Composite, useCompositeStoreV2: useCompositeStore } =
-	unlock( componentsPrivateApis );
 const noop = () => {};
+
+const useCompositeStore = Composite.useStore;
 
 function DownloadableBlocksList( { items, onHover = noop, onSelect } ) {
 	const composite = useCompositeStore();
@@ -26,7 +25,7 @@ function DownloadableBlocksList( { items, onHover = noop, onSelect } ) {
 	}
 
 	return (
-		<Composite
+		<Composite.Root
 			store={ composite }
 			role="listbox"
 			className="block-directory-downloadable-blocks-list"
@@ -57,7 +56,7 @@ function DownloadableBlocksList( { items, onHover = noop, onSelect } ) {
 					/>
 				);
 			} ) }
-		</Composite>
+		</Composite.Root>
 	);
 }