+++ packages/block-editor/src/components/block-switcher/pattern-transformations-menu.js
@@ -7,11 +7,11 @@ import { useInstanceId } from '@wordpress/compose';
 import { chevronRight } from '@wordpress/icons';
 
 import {
+	Composite,
 	MenuGroup,
 	MenuItem,
 	Popover,
 	VisuallyHidden,
-	privateApis as componentsPrivateApis,
 } from '@wordpress/components';
 
 /**
@@ -19,13 +19,8 @@ import {
  */
 import BlockPreview from '../block-preview';
 import useTransformedPatterns from './use-transformed-patterns';
-import { unlock } from '../../lock-unlock';
 
-const {
-	CompositeV2: Composite,
-	CompositeItemV2: CompositeItem,
-	useCompositeStoreV2: useCompositeStore,
-} = unlock( componentsPrivateApis );
+const useCompositeStore = Composite.useStore;
 
 function PatternTransformationsMenu( {
 	blocks,
@@ -82,7 +77,7 @@ function PreviewPatternsPopover( { patterns, onSelect } ) {
 function BlockPatternsList( { patterns, onSelect } ) {
 	const composite = useCompositeStore();
 	return (
-		<Composite
+		<Composite.Root
 			store={ composite }
 			role="listbox"
 			className="block-editor-block-switcher__preview-patterns-container"
@@ -95,7 +90,7 @@ function BlockPatternsList( { patterns, onSelect } ) {
 					onSelect={ onSelect }
 				/>
 			) ) }
-		</Composite>
+		</Composite.Root>
 	);
 }
 
@@ -109,7 +104,7 @@ function BlockPattern( { pattern, onSelect } ) {
 	);
 	return (
 		<div className={ `${ baseClassName }-list__list-item` }>
-			<CompositeItem
+			<Composite.Item
 				render={
 					<div
 						role="option"
@@ -129,7 +124,7 @@ function BlockPattern( { pattern, onSelect } ) {
 				<div className={ `${ baseClassName }-list__item-title` }>
 					{ pattern.title }
 				</div>
-			</CompositeItem>
+			</Composite.Item>
 			{ !! pattern.description && (
 				<VisuallyHidden id={ descriptionId }>
 					{ pattern.description }